# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Songtai Life Partners Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type PartnerStatus = typeof PartnerStatus[keyof typeof PartnerStatus];


export const PartnerStatus = {
  pending: 'pending',
  active: 'active',
  suspended: 'suspended',
} as const;

export interface Partner {
  id: string;
  slug: string;
  status: PartnerStatus;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  heroTitleEn?: string | null;
  /** @nullable */
  heroTitleFr?: string | null;
  /** @nullable */
  heroSubtitleEn?: string | null;
  /** @nullable */
  heroSubtitleFr?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  pendingContactName?: string | null;
  /** @nullable */
  pendingContactPhone?: string | null;
  createdAt: string;
}

export interface PartnerInput {
  /**
     * @minLength 2
     * @maxLength 60
     */
  slug: string;
  whatsappNumber?: string;
  contactEmail?: string;
  heroTitleEn?: string;
  heroTitleFr?: string;
  heroSubtitleEn?: string;
  heroSubtitleFr?: string;
  heroImageUrl?: string;
  pendingContactName?: string;
  pendingContactPhone?: string;
}

export interface PartnerUpdate {
  whatsappNumber?: string;
  contactEmail?: string;
  heroTitleEn?: string;
  heroTitleFr?: string;
  heroSubtitleEn?: string;
  heroSubtitleFr?: string;
  heroImageUrl?: string;
}

export type PartnerStatusUpdateStatus = typeof PartnerStatusUpdateStatus[keyof typeof PartnerStatusUpdateStatus];


export const PartnerStatusUpdateStatus = {
  pending: 'pending',
  active: 'active',
  suspended: 'suspended',
} as const;

export interface PartnerStatusUpdate {
  status: PartnerStatusUpdateStatus;
}

export interface PartnerStats {
  total: number;
  active: number;
  pending: number;
  suspended: number;
}

export interface Product {
  id: string;
  nameEn: string;
  nameFr: string;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  descriptionFr?: string | null;
  priceXaf: number;
  pvPoints: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  stock?: number;
  isActive: boolean;
}

export interface ProductInput {
  nameEn: string;
  nameFr: string;
  descriptionEn?: string;
  descriptionFr?: string;
  priceXaf: number;
  pvPoints: number;
  category: string;
  imageUrl?: string;
  stock?: number;
}

export interface ProductUpdate {
  nameEn?: string;
  nameFr?: string;
  descriptionEn?: string;
  descriptionFr?: string;
  priceXaf?: number;
  pvPoints?: number;
  category?: string;
  imageUrl?: string;
  stock?: number;
  isActive?: boolean;
}

export interface Testimonial {
  id: string;
  authorName: string;
  /** @nullable */
  authorRole?: string | null;
  contentEn: string;
  contentFr: string;
  rating: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt?: string;
}

export interface TestimonialInput {
  authorName: string;
  authorRole?: string;
  contentEn: string;
  contentFr: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  imageUrl?: string;
}

export interface GalleryImage {
  id: string;
  imageUrl: string;
  /** @nullable */
  captionEn?: string | null;
  /** @nullable */
  captionFr?: string | null;
  sortOrder: number;
}

export interface GalleryImageInput {
  imageUrl: string;
  captionEn?: string;
  captionFr?: string;
  sortOrder?: number;
}

export interface FaqItem {
  id: string;
  questionEn: string;
  questionFr: string;
  answerEn: string;
  answerFr: string;
  /** @nullable */
  category?: string | null;
  sortOrder: number;
}

export interface FaqItemInput {
  questionEn: string;
  questionFr: string;
  answerEn: string;
  answerFr: string;
  category?: string;
  sortOrder?: number;
}

export interface AboutContent {
  id: string;
  storyEn: string;
  storyFr: string;
  missionEn: string;
  missionFr: string;
  visionEn: string;
  visionFr: string;
  /** @nullable */
  imageUrl?: string | null;
  updatedAt?: string;
}

export interface AboutContentInput {
  storyEn?: string;
  storyFr?: string;
  missionEn?: string;
  missionFr?: string;
  visionEn?: string;
  visionFr?: string;
  imageUrl?: string;
}