# lib/api-zod/src/generated/types/aboutContentInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Songtai Life Partners Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AboutContentInput {
  storyEn?: string;
  storyFr?: string;
  missionEn?: string;
  missionFr?: string;
  visionEn?: string;
  visionFr?: string;
  imageUrl?: string;
}