# lib/api-zod/src/generated/types/partnerStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Songtai Life Partners Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PartnerStats {
  total: number;
  active: number;
  pending: number;
  suspended: number;
}